+++ lib/src/controller.dart
@@ -8,25 +8,29 @@ class CropController {
   set delegate(CropControllerDelegate value) => _delegate = value;
 
   /// crop given image with current configuration
-  void crop() => _delegate.onCrop();
+  void crop() => _delegate.onCrop(false);
+
+  /// crop given image with current configuration and circle shape.
+  void cropCircle() => _delegate.onCrop(true);
 
   /// change fixed aspect ratio
   /// if [value] is null, cropping area can be moved without fixed aspect ratio.
   set aspectRatio(double? value) => _delegate.onChangeAspectRatio(value);
 
-  /// change if cropping with circle shape.
+  /// change if cropping with circle shaped UI.
   /// if [value] is true, [aspectRatio] automatically fixed with 1
-  set isCircle(bool value) => _delegate.onChangeIsCircle(value);
+  set withCircleUi(bool value) => _delegate.onChangeWithCircleUi(value);
 }
 
 /// Delegate of actions from [CropController]
 class CropControllerDelegate {
   /// callback that [CropController.crop] is called.
-  late VoidCallback onCrop;
+  /// the meaning of the value is if cropping a image with circle shape.
+  late ValueChanged<bool> onCrop;
 
   /// callback that [CropController.aspectRatio] is set.
   late ValueChanged<double?> onChangeAspectRatio;
 
-  /// callback that [CropController.isCircle] is changed.
-  late ValueChanged<bool> onChangeIsCircle;
+  /// callback that [CropController.withCircleUi] is changed.
+  late ValueChanged<bool> onChangeWithCircleUi;
 }